build: run the cmake build/test as part of make distcheck
This is useful for ensuring that the cmake build works even from a
source tarball generated by automake.  Also good smoke-test for ensuring
all required files are bundled.

Here I fixed a couple small issues with the cmake build that only came
up in the corner case of trying to run it within a `make distcheck`
build directory.

Also deleted now redundant copies of sae_par.h and mers.h which are
already available in `ast_tester/star_compat/`
