--- cmake/sae_par.h
@@ -1,11 +1,0 @@
-/*
- *  Minimal stub for the Starlink SAE_PAR header.
- *  Provides only the status constants needed by the AST test suite.
- */
-#ifndef SAE_PAR_H_INCLUDED
-#define SAE_PAR_H_INCLUDED
-
-#define SAI__OK    0
-#define SAI__ERROR 1
-
-#endif
